refactor: DTO 패키지 구조 변경
request와 response로 분리
