+++ back/babble/src/main/java/gg/babble/babble/controller/GameController.java
@@ -1,7 +1,7 @@
 package gg.babble.babble.controller;
 
-import gg.babble.babble.dto.GameImageResponse;
-import gg.babble.babble.dto.IndexPageGameResponse;
+import gg.babble.babble.dto.response.GameImageResponse;
+import gg.babble.babble.dto.response.IndexPageGameResponse;
 import gg.babble.babble.service.GameService;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
