+++ back/babble/src/main/java/gg/babble/babble/dto/request/MessageRequest.java
@@ -1,6 +1,6 @@
-package gg.babble.babble.dto;
+package gg.babble.babble.dto.request;
 
-import javax.validation.constraints.NotBlank;
+
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
@@ -12,12 +12,12 @@ import lombok.Setter;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class UserJoinRequest {
+public class MessageRequest {
 
     @NotNull(message = "userId는 비어있을 수 없습니다.")
     @Positive(message = "userId는 음수일 수 없습니다.")
     private Long userId;
 
-    @NotBlank(message = "sessionId는 비어있을 수 없습니다.")
-    private String sessionId;
+    @NotNull(message = "채팅 메시지는 Null 일 수 없습니다.")
+    private String content;
 }
