+++ back/babble/src/main/java/gg/babble/babble/dto/response/TagResponse.java
@@ -1,4 +1,4 @@
-package gg.babble.babble.dto;
+package gg.babble.babble.dto.response;
 
 import gg.babble.babble.domain.tag.Tag;
 import lombok.AllArgsConstructor;
