+++ back/babble/src/main/java/gg/babble/babble/dto/response/UserListUpdateResponse.java
@@ -1,4 +1,4 @@
-package gg.babble.babble.dto;
+package gg.babble.babble.dto.response;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
