+++ back/babble/src/main/java/gg/babble/babble/service/GameService.java
@@ -3,8 +3,8 @@ package gg.babble.babble.service;
 import gg.babble.babble.domain.Game;
 import gg.babble.babble.domain.Games;
 import gg.babble.babble.domain.repository.GameRepository;
-import gg.babble.babble.dto.GameImageResponse;
-import gg.babble.babble.dto.IndexPageGameResponse;
+import gg.babble.babble.dto.response.GameImageResponse;
+import gg.babble.babble.dto.response.IndexPageGameResponse;
 import gg.babble.babble.exception.BabbleNotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
