+++ back/babble/src/main/java/gg/babble/babble/service/RoomService.java
@@ -4,12 +4,12 @@ import gg.babble.babble.domain.repository.RoomRepository;
 import gg.babble.babble.domain.room.MaxHeadCount;
 import gg.babble.babble.domain.room.Room;
 import gg.babble.babble.domain.user.User;
-import gg.babble.babble.dto.CreatedRoomResponse;
-import gg.babble.babble.dto.FoundRoomResponse;
-import gg.babble.babble.dto.RoomRequest;
-import gg.babble.babble.dto.UserJoinRequest;
-import gg.babble.babble.dto.UserListUpdateResponse;
-import gg.babble.babble.dto.UserResponse;
+import gg.babble.babble.dto.request.RoomRequest;
+import gg.babble.babble.dto.request.UserJoinRequest;
+import gg.babble.babble.dto.response.CreatedRoomResponse;
+import gg.babble.babble.dto.response.FoundRoomResponse;
+import gg.babble.babble.dto.response.UserListUpdateResponse;
+import gg.babble.babble.dto.response.UserResponse;
 import gg.babble.babble.exception.BabbleNotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
