+++ back/babble/src/main/java/gg/babble/babble/service/TagService.java
@@ -2,8 +2,8 @@ package gg.babble.babble.service;
 
 import gg.babble.babble.domain.repository.TagRepository;
 import gg.babble.babble.domain.tag.Tag;
-import gg.babble.babble.dto.TagRequest;
-import gg.babble.babble.dto.TagResponse;
+import gg.babble.babble.dto.request.TagRequest;
+import gg.babble.babble.dto.response.TagResponse;
 import gg.babble.babble.exception.BabbleNotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
