+++ back/babble/src/test/java/gg/babble/babble/service/GameServiceTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import gg.babble.babble.ApplicationTest;
-import gg.babble.babble.dto.GameImageResponse;
-import gg.babble.babble.dto.IndexPageGameResponse;
+import gg.babble.babble.dto.response.GameImageResponse;
+import gg.babble.babble.dto.response.IndexPageGameResponse;
 import gg.babble.babble.exception.BabbleNotFoundException;
 import java.util.Arrays;
 import java.util.List;
