+++ back/babble/src/test/java/gg/babble/babble/service/RoomServiceTest.java
@@ -7,15 +7,15 @@ import gg.babble.babble.ApplicationTest;
 import gg.babble.babble.domain.Game;
 import gg.babble.babble.domain.tag.Tag;
 import gg.babble.babble.domain.user.User;
-import gg.babble.babble.dto.CreatedRoomResponse;
-import gg.babble.babble.dto.FoundRoomResponse;
-import gg.babble.babble.dto.GameResponse;
-import gg.babble.babble.dto.HeadCountResponse;
-import gg.babble.babble.dto.RoomRequest;
-import gg.babble.babble.dto.TagRequest;
-import gg.babble.babble.dto.TagResponse;
-import gg.babble.babble.dto.UserJoinRequest;
-import gg.babble.babble.dto.UserResponse;
+import gg.babble.babble.dto.request.RoomRequest;
+import gg.babble.babble.dto.request.TagRequest;
+import gg.babble.babble.dto.request.UserJoinRequest;
+import gg.babble.babble.dto.response.CreatedRoomResponse;
+import gg.babble.babble.dto.response.FoundRoomResponse;
+import gg.babble.babble.dto.response.GameResponse;
+import gg.babble.babble.dto.response.HeadCountResponse;
+import gg.babble.babble.dto.response.TagResponse;
+import gg.babble.babble.dto.response.UserResponse;
 import gg.babble.babble.exception.BabbleNotFoundException;
 import java.util.ArrayList;
 import java.util.Arrays;
